backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.1
mysqlclient==2.2.7
pymongo==4.14.1
python-decouple==3.8
